converting strings to floats

+++ src/pandapipes/converter/stanet/table_creation.py
@@ -76,7 +76,7 @@ def create_junctions_from_nodes(net, stored_data, net_params, index_mapping, add
         add_info["stanet_layer"] = node_table.LAYER.values.astype(str)
     temperatures = pd.Series(net_params["medium_temp_K"], index=node_table.index, dtype=np.float64)
     eg_ind = ((node_table.FSTATUS == '?') & (node_table.DSTATUS == '!')).values
-    eg_temps = node_table.loc[eg_ind, "TMESS"].values + 273.15
+    eg_temps = node_table.loc[eg_ind, "TMESS"].values.astype(float) + 273.15
     if net_params["calculate_temp"]:
         temperatures.loc[eg_ind] = eg_temps
     else:
@@ -93,7 +93,7 @@ def create_junctions_from_nodes(net, stored_data, net_params, index_mapping, add
         stanet_active=node_table.ISACTIVE.values.astype(np.bool_),
         stanet_system=CLIENT_TYPES_OF_PIPES[MAIN_PIPE_TYPE], **add_info)
     for eg_junc, p_bar, t_k in zip(junction_indices[eg_ind], eg_press, eg_temps):
-        pandapipes.create_ext_grid(net, eg_junc, p_bar, t_k, type="pt",
+        pandapipes.create_ext_grid(net, eg_junc, float(p_bar), float(t_k), type="pt",
                                    stanet_system=CLIENT_TYPES_OF_PIPES[MAIN_PIPE_TYPE])
     index_mapping["nodes"] = dict(zip(stanet_nrs, junction_indices))
 
@@ -250,7 +250,7 @@ def create_slider_valves(net, stored_data, index_mapping, add_layers,
                            f"The diameter will be set to 1 m.")
             slider_valves.loc[slider_valves.DM == 0, 'DM'] = 1e3
         pandapipes.create_valves(
-            net, from_junctions, to_junctions, slider_valves.DM.values / 1000,
+            net, from_junctions, to_junctions, slider_valves.DM.values.astype(float) / 1000,
             opened=slider_valves.TYP.astype(np.int32).replace(opened_types).values,
             loss_coefficient=slider_valves.ZETA.values, name=slider_valves.STANETID.values,
             type="slider_valve_" + valve_system,
@@ -389,10 +389,8 @@ def create_control_components(net, stored_data, index_mapping, net_params, add_l
             stanet_is_closed=fully_closed[is_pc],
             stanet_flow_kgps=flow[is_pc],
             stanet_active=control_table.ISACTIVE.values[is_pc].astype(np.bool_),
-            max_mdot_kg_per_s=control_table.QSOLL.values[is_pc] / 3600 * fluid.get_density(NORMAL_TEMPERATURE)
-
-            **add_info
-        )
+            max_mdot_kg_per_s=control_table.QSOLL.values[is_pc].astype(float) / 3600 * fluid.get_density(NORMAL_TEMPERATURE),
+            **add_info)
 
         drop_eg = net.ext_grid.loc[net.ext_grid.junction.isin(to_junctions[is_pc])].index
         net.ext_grid.drop(drop_eg, inplace=True)
@@ -577,15 +575,15 @@ def create_pipes_from_connections(net, stored_data, connection_table, index_mapp
                          pipe_data.loc[pipe_nums, "ANFNR"].values[previous_different])
     con_to = np.insert(cons.RECNO.values, next_different_num + 1,
                        pipe_data.loc[pipe_nums, "ENDNR"].values[next_different])
-    vm_from = np.insert(cons.VMB.values, next_different_num, cons.VMA.values[next_different_num])
-    vm_to = np.insert(cons.VMA.values, next_different_num, cons.VMB.values[next_different_num])
+    vm_from = np.insert(cons.VMB.values.astype(float), next_different_num, cons.VMA.values[next_different_num].astype(float),)
+    vm_to = np.insert(cons.VMA.values.astype(float), next_different_num, cons.VMB.values[next_different_num].astype(float),)
     vm = (vm_from + vm_to) / 2
 
     pipe_sections = pd.DataFrame({
         "SNUM": pipe_numbers, "rel_length": rel_lengths, "start_pos": start_pos, "end_pos": end_pos,
         "from_type": type_from, "to_type": type_to, "from_node": con_from, "to_node": con_to,
         "full_geo": pipe_geodata.loc[pipe_numbers], "vm": vm,
-        "length": rel_lengths * pipe_data.RORL.loc[pipe_numbers].values,
+        "length": rel_lengths * pipe_data.RORL.loc[pipe_numbers].values.astype(float),
         "aux": np.ones(len(pipe_numbers), dtype=np.int32)
     })
     pipe_sections["section_no"] = pipe_sections.groupby("SNUM").aux.cumsum()
@@ -614,7 +612,7 @@ def create_pipes_from_connections(net, stored_data, connection_table, index_mapp
         alpha = pipes.WDZAHL.values.astype(np.float64)
     pandapipes.create_pipes_from_parameters(
         net, pipe_sections.fj.values, pipe_sections.tj.values, pipe_sections.length.values / 1000,
-        pipes.DM.values / 1000, pipes.RAU.values, pipes.ZETA.values, type="main_pipe",
+        pipes.DM.values.astype(float) / 1000, pipes.RAU.values.astype(float), pipes.ZETA.values.astype(float), type="main_pipe",
         stanet_std_type=pipes.ROHRTYP.values, in_service=pipes.ISACTIVE.values, text_k=text_k,
         alpha_w_per_m2k=alpha,
         name=["pipe_%s_%s_%s" % (nf, nt, sec) for nf, nt, sec in zip(
